lmao it's the network interface.
originally, sodium lived in a Dell computer. I sold that one since, and
bought a new computer. this makes me fairly certain vpn.mod.nix is older
than january, likely dating back to fall of 2023. but given that sodium
had moved homes, the ethernet interface was changed. Of course
redirect-gateway didn't work; the iptables command was trying to route
to a nonexistent device!