fix Steam build broken by https://github.com/NixOS/nixpkgs/pull/349109
note: if you came here from nixpkgs (i.e. the backlink in that PR), this
isn't really a mistake on nixpkgs' part, i was depending on internal
details of the package. it is still, nonetheless, a breakage caused by
that PR. i'm pretty sure that's the PR that caused it to break, at
least.