enable flashrom support for fwupd
required to upgrade firmware on nitrogen, which is a StarLite 5
found out that flashrom plugin is required from
https://github.com/StarLabsLtd/firmware/issues/159
has been an open issue for 2.5 years on nixpkgs
https://github.com/NixOS/nixpkgs/issues/153238

this doesn't work as-is with the nixpkgs flashrom package, because it's
not built using meson i guess?
i found out abouut the flashrom-meson due to the backlink that this
issue caused: https://github.com/Roger/flashrom-meson-nix/issues/1
and if you read this commit coming from that issue, the answer is that
yeah. it does work, kind of.

after the changes in this commit, i was able to update my bios but not
my embedded controller (EC). so bios is now on 24.09, but EC is on
24.06. EC is not detected as a device by fwupd.

new firmware also fixed the tpm timeout causing my boot to take
excruciatingly long