make sops reference the secrets.yaml directly
this removes an annoying source of rebuild noise: previously, any change
to a dirty tree will cause the base "${self}" to change and therefore
change the entire sops config closure. now, when referencing the file
directly, it is copied by itself to the nix store, and its hash does not
depend on any other part of the repo.